refactor: types and getters to modules and remove redundant core function
